ResumeScreening/AI_Service/app.py
import streamlit as st
from PyPDF2 import PdfReader
import spacy
import en_core_web_sm
import requests
import google.generativeai as genai
import base64

# Load the spaCy NLP model
nlp = spacy.load("en_core_web_sm")

# Gemini API key (you should set this as an environment variable)
API_KEY = "AIzaSyClYAGCv1b4evmni_sBBq0SeuhC4qeVqCM"
genai.configure(api_key=API_KEY)
# Function to extract text from PDF
def get_gemini_response(input, pdf_content, prompt):
    model = genai.GenerativeModel('gemini-1.5-flash')
    response = model.generate_content([input, pdf_content[0], prompt])
    return response.text


def input_pdf_setup(uploaded_file):
    if uploaded_file is not None:
        reader = PdfReader(uploaded_file)
        pdf_text = ""

        # Extract text from all pages
        for page_num in range(len(reader.pages)):
            pdf_text += reader.pages[page_num].extract_text()

        pdf_parts = [
            {
                "mime_type": "text/plain",
                "data": base64.b64encode(pdf_text.encode()).decode()  # encode to base64
            }
        ]
        return pdf_parts
    else:
        raise FileNotFoundError("No file uploaded")


## Streamlit App

st.set_page_config(page_title="ATS Resume Expert")
st.header("ATS Tracking System")
input_text = st.text_area("Job Description: ", key="input")
uploaded_file = st.file_uploader("Upload your resume (PDF)...", type=["pdf"])

if uploaded_file is not None:
    st.write("PDF Uploaded Successfully")

submit1 = st.button("Tell Me About the Resume")
submit3 = st.button("Percentage match")

input_prompt1 = """
You are an experienced Technical Human Resource Manager, your task is to review the provided resume against the job description. 
Please share your professional evaluation on whether the candidate's profile aligns with the role. 
Highlight the strengths and weaknesses of the applicant in relation to the specified job requirements.
"""

input_prompt3 = """
You are a skilled ATS (Applicant Tracking System) scanner with a deep understanding of data science and ATS functionality, 
your task is to evaluate the resume against the provided job description. Give me the percentage match if the resume matches
the job description. First, the output should come as percentage and then keywords missing and finally, final thoughts.
"""

if submit1:
    if uploaded_file is not None:
        pdf_content = input_pdf_setup(uploaded_file)
        response = get_gemini_response(input_prompt1, pdf_content, input_text)
        st.subheader("The Response is")
        st.write(response)
    else:
        st.write("Please upload the resume")

elif submit3:
    if uploaded_file is not None:
        pdf_content = input_pdf_setup(uploaded_file)
        response = get_gemini_response(input_prompt3, pdf_content, input_text)
        st.subheader("The Response is")
        st.write(response)
    else:
        st.write("Please upload the resume")